Fix compatibility with old .file-hashes files
boto3 uploads /some/key to bucket//some/key, leaving part of the path
with "//", or a folder without a name. To remedy this, I had lstrip'd
the / from the key before upload. I didn't think about the fact that
boto2 works fine with this, and the old .file-hashes file would have the
keys with the prefixing forward slash. This solves that